+++ src/app/user/faq/page.tsx
@@ -1,6 +1,5 @@
 import { FAQSupportSection } from "../../../components/features/user/faq/FAQSupportSection";
-import Faq from '../../../components/features/user/faq/faq/faq'
-
+import Faq from "../../../components/features/user/faq/faq";
 
 export default function FaqPage() {
   return (
@@ -10,7 +9,8 @@ export default function FaqPage() {
           <div className="flex-grow">
             <h1 className="text-3xl font-bold">Frequently Asked Questions</h1>
             <p className="text-gray-300">
-              Find answers to common questions below. If you need further assistance, our support team is here to help.
+              Find answers to common questions below. If you need further
+              assistance, our support team is here to help.
             </p>
           </div>
           <Faq />
